backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24.0,<0.26.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
openai==1.3.7
azure-cognitiveservices-speech==1.34.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
pillow==10.1.0
aiofiles==23.2.1
boto3==1.34.0
supabase==2.0.2
